(PC-26667) refactor(apiHelpers): Move remaining lifetime helper
