+++ src/api/apiHelpers.ts
@@ -6,7 +6,7 @@ import { getCodePushId } from 'api/getCodePushId'
 import { navigateFromRef } from 'features/navigation/navigationRef'
 import { env } from 'libs/environment'
 import { Headers } from 'libs/fetch'
-import { decodeToken, getTokenStatus } from 'libs/jwt'
+import { computeTokenRemainingLifetimeInMs, getTokenStatus } from 'libs/jwt'
 import { clearRefreshToken, getRefreshToken } from 'libs/keychain'
 import { eventMonitoring } from 'libs/monitoring'
 import { getAppVersion } from 'libs/packageJson'
@@ -201,19 +201,6 @@ const extractResponseBody = async (response: Response): Promise<string> => {
     return response.json()
   }
   return response.text()
-}
-
-export const computeTokenRemainingLifetimeInMs = (encodedToken: string): number | undefined => {
-  const token = decodeToken(encodedToken)
-
-  if (token) {
-    const tokenExpirationInMs = token.exp * 1000
-    const currentDateInMs = Date.now()
-    const lifetimeInMs = tokenExpirationInMs - currentDateInMs
-    return lifetimeInMs
-  }
-
-  return undefined
 }
 
 // In this case, the following `any` is not that much of a problem in the context of usage
