+++ src/features/auth/context/AuthContext.tsx
@@ -3,7 +3,6 @@ import React, { memo, useCallback, useContext, useEffect, useMemo, useRef, useSt
 import { QueryObserverResult } from 'react-query'
 
 import { api } from 'api/api'
-import { computeTokenRemainingLifetimeInMs } from 'api/apiHelpers'
 import { UserProfileResponse } from 'api/gen'
 import { useCookies } from 'features/cookies/helpers/useCookies'
 // eslint-disable-next-line no-restricted-imports
@@ -11,7 +10,11 @@ import { amplitude } from 'libs/amplitude'
 import { useAppStateChange } from 'libs/appState'
 // eslint-disable-next-line no-restricted-imports
 import { firebaseAnalytics } from 'libs/firebase/analytics'
-import { getTokenStatus, getUserIdFromAccessToken } from 'libs/jwt'
+import {
+  computeTokenRemainingLifetimeInMs,
+  getTokenStatus,
+  getUserIdFromAccessToken,
+} from 'libs/jwt'
 import { getTokenExpirationDate } from 'libs/jwt/getTokenExpirationDate'
 import { getRefreshToken } from 'libs/keychain'
 import { eventMonitoring } from 'libs/monitoring'
